subprocess agregado al codigo para manejar los errores

+++ build.py
@@ -3,9 +3,11 @@ from datetime import datetime
 from gitHandler import git_file_status, git_commit
 from jsonHandler import write_to_child, get_top_level_keys, read_from_child, sort_json
 from reducer import reducer
+from smdbHandler import get_all_smdb
 import csv
 import os
 import shutil
+import subprocess
 
 
 def load_config(_section="build_reduced", _file="config.ini"):
@@ -16,19 +18,19 @@ def load_config(_section="build_reduced", _file="config.ini"):
 
 
 def select_database(_path, _sha1):
-    smdb, basename = get_missing_smdb(_sha1)
+    smdb, basename = missing_smdb(_sha1)
 
     [print(index, value) for index, value in enumerate(basename)]
     print("999 for all SMDB files")
     index = input("Select one SMDB file: ")
 
     if index == "999":
-        return get_all_smdb(_path)
+        return all_smdb(_path)
 
     return smdb[int(index)], basename[int(index)]
 
 
-def get_missing_smdb(_sha1):
+def missing_smdb(_sha1):
     db = get_top_level_keys()
     missing = [x for x in db if read_from_child(x, "missing")]
     not_verified = [
@@ -43,11 +45,10 @@ def get_missing_smdb(_sha1):
     return smdb, basename
 
 
-def get_all_smdb(_path):
+def all_smdb(_path):
     os.system("clear" if os.name == "posix" else "cls")
 
-    smdb = sorted([x for x in os.listdir(_path) if x.endswith(".txt")], key=str.lower)
-    basename = [os.path.splitext(os.path.basename(x))[0] for x in smdb]
+    smdb, basename = get_all_smdb(_path)
 
     [print(index, value) for index, value in enumerate(basename)]
     index = input("Select one SMDB file: ")
@@ -56,7 +57,23 @@ def get_all_smdb(_path):
 
 
 def uncompress_7z(_file, _output_folder):
-    os.system(f"7z x '{_file}' -o'{_output_folder}' -y")
+    uncompress = [
+        "7z",
+        "x",
+        _file,
+        f"-o{_output_folder}",
+        "-y",
+        "-aoa",
+    ]
+
+    try:
+        subprocess.run(uncompress, check=True)
+    except subprocess.CalledProcessError as e:
+        print(f"Error al ejecutar el comando: {e.returncode}")
+        exit()
+    except Exception as e:
+        print(f"Ocurrió un error: {e}")
+        exit()
 
 
 def compress_7z(_basename, _path):
@@ -105,7 +122,25 @@ def build_from_main(_folder, _romimport):
         print(f"Main folder {_folder} is empty")
         return False
 
-    shutil.move(_folder, _romimport)
+    if not os.path.exists(_romimport):
+        shutil.move(_folder, _romimport)
+        return True
+
+    for root, _, files in os.walk(_folder):
+        relative_path = os.path.relpath(root, _folder)
+        romimport_dir = os.path.join(_romimport, relative_path)
+
+        if not os.path.exists(romimport_dir):
+            input(f"romimport_dir: {romimport_dir}")
+            os.makedirs(romimport_dir)
+
+        for file in files:
+            folder_file = os.path.join(root, file)
+            romimport_file = os.path.join(romimport_dir, file)
+            if not os.path.exists(romimport_file):
+                shutil.move(folder_file, romimport_file)
+
+    shutil.rmtree(_folder)
 
     return True
 
@@ -127,9 +162,31 @@ def build():
         if not build_from_masters(masters, basename, romimport):
             return print("No files to build")
 
-    build = f"python {section['script']} --input_folder '{romimport}' --database '{smdb}' --output_folder '{folder}' --missing '{missing}' --skip_existing --drop_initial_directory --file_strategy hardlink"
+    build = [
+        "python",
+        section["script"],
+        "--input_folder",
+        romimport,
+        "--database",
+        smdb,
+        "--output_folder",
+        folder,
+        "--missing",
+        missing,
+        "--skip_existing",
+        "--drop_initial_directory",
+        "--file_strategy",
+        "hardlink",
+    ]
 
-    os.system(build)
+    try:
+        subprocess.run(build, check=True)
+    except subprocess.CalledProcessError as e:
+        print(f"Error al ejecutar el comando: {e.returncode}")
+        exit()
+    except Exception as e:
+        print(f"Ocurrió un error: {e}")
+        exit()
 
     shutil.rmtree(romimport)
